[Nexus] moved IHierarchicalDataModel to its own plug-in project, org.eclipse.dawnsci.hdf5.model
