+++ org.eclipse.dawnsci.hdf5.model/.gitignore
@@ -1,0 +1,1 @@
+/bin/
