+++ org.eclipse.dawnsci.hdf5.model/META-INF/MANIFEST.MF
@@ -1,0 +1,13 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: HDF5 Hierarchical Model
+Bundle-SymbolicName: org.eclipse.dawnsci.hdf5.model
+Bundle-Version: 1.0.0.qualifier
+Bundle-Activator: org.eclipse.dawnsci.hdf5.model.Activator
+Bundle-RequiredExecutionEnvironment: JavaSE-1.8
+Export-Package: org.eclipse.dawnsci.hdf5.model,
+ org.eclipse.dawnsci.hdf5.model.internal;x-internal:=true
+Require-Bundle: org.eclipse.core.runtime;bundle-version="3.8.0",
+ org.eclipse.core.resources;bundle-version="3.8.1",
+ org.eclipse.dawnsci.hdf5;bundle-version="1.3.0",
+ ncsa.hdf;bundle-version="[2.9.0,3.0.0)"
