+++ org.eclipse.dawnsci.hdf5.model/src/org/eclipse/dawnsci/hdf5/model/IHierarchicalDataFileModel.java
