+++ org.eclipse.dawnsci.hdf5/META-INF/MANIFEST.MF
@@ -3,19 +3,14 @@ Bundle-ManifestVersion: 2
 Bundle-Name: Hdf5
 Bundle-SymbolicName: org.eclipse.dawnsci.hdf5;singleton:=true
 Bundle-Version: 1.3.0.qualifier
-Bundle-Activator: org.eclipse.dawnsci.hdf5.Activator
 Bundle-Vendor: ESRF
-Require-Bundle: org.eclipse.core.runtime,
- ncsa.hdf;bundle-version="[2.9.0,3.0.0)",
- org.eclipse.core.resources;bundle-version="3.8.1",
+Require-Bundle: ncsa.hdf;bundle-version="[2.9.0,3.0.0)",
  org.eclipse.dawnsci.analysis.api;bundle-version="1.0.0",
  org.eclipse.dawnsci.analysis.dataset;bundle-version="1.0.0"
 Bundle-RequiredExecutionEnvironment: JavaSE-1.7
 Bundle-ActivationPolicy: lazy
 Bundle-ClassPath: .
 Export-Package: org.eclipse.dawnsci.hdf5,
- org.eclipse.dawnsci.hdf5.model,
- org.eclipse.dawnsci.hdf5.model.internal;x-internal:=true,
  org.eclipse.dawnsci.hdf5.nexus,
  org.eclipse.dawnsci.hdf5.operation
 Import-Package: org.slf4j;version="[1.7.6,1.7.7)"
